+++ scripts/core.js
@@ -1,90 +1,239 @@
-import {
-  name,
-  templates,
-  log
-} from './config.js'
+import { name, templates, log } from "./config.js";
+import Setting from "./setting.js";
 
 export default class Core extends FormApplication {
+  /**
+   * @param {Array.<Object>} settings Read from previously exported settings
+   */
+  constructor(settings) {
+    super();
+    this.settings = [];
+    this.hasWorldSettings = false;
+    this.playerSettings = [];
+    this.hasPlayerSettings = false;
+    this.notChangedSettings = [];
+    this.notChangedPlayers = [];
+    this.notFoundPlayers = [];
+
+    if (settings && Array.isArray(settings)) {
+      log(false, "Parsing provided settings", settings);
+
+      settings.forEach((data) => {
+        try {
+          let setting = new Setting(data);
+          if (setting) {
+            switch (setting.type) {
+              case Setting.WorldType:
+                if (setting.hasChanges()) {
+                  this.settings.push(setting.value);
+                  this.hasWorldSettings = true;
+                } else {
+                  this.notChangedSettings.push(setting.data.key);
+                }
+                break;
+              case Setting.PlayerType:
+                if (!setting.hasChanges()) {
+                  this.notChangedPlayers.push(setting.data.name);
+                  break;
+                }
+                if (setting.value.playerNotFound) {
+                  this.notFoundPlayers.push(setting.value);
+                  break;
+                }
+                this.playerSettings.push(setting.value);
+                this.hasPlayerSettings = true;
+            }
+          }
+        } catch (e) {
+          log(false, "Error importing setting:", e, data);
+        }
+      });
+    }
+
+    log(false, "Processing world settings", this.settings);
+    log(false, "Processing player settings", this.playerSettings);
+  }
 
   static get defaultOptions() {
     return mergeObject(super.defaultOptions, {
-      classes: ['copy-environment-world-settings'],
-      height: 'auto',
-      width: 600,
-      id: `${name}-world-settings`,
-      title: `${name}.forms.world.title`,
-      template: templates.worldSettings,
-      tabs: [{
-        navSelector: '.tabs',
-        contentSelector: 'form',
-        initial: 'settings',
-      }]
+      classes: ["copy-environment-settings"],
+      height: "auto",
+      width: Math.ceil(window.innerWidth / 2),
+      id: `${name}-settings`,
+      title: `${name}.title`,
+      template: templates.settings,
     });
   }
 
   getData() {
-    return {};
+    return {
+      settings: this.settings,
+      playerSettings: this.playerSettings,
+      hasWorldSettings: this.hasWorldSettings,
+      hasPlayerSettings: this.hasPlayerSettings,
+      hasChanges: this.hasWorldSettings || this.hasPlayerSettings,
+      notChangedSettings: this.notChangedSettings,
+      notChangedPlayers: this.notChangedPlayers,
+      notFoundPlayers: this.notFoundPlayers,
+    };
   }
-
 
   activateListeners(html) {
     super.activateListeners(html);
 
-    html.on('click', '.export-save-world-env', () => {
-      Core.saveSummaryAsJSON();
+    html.on("click", ".close", () => {
       this.close();
     });
 
-    html.on('click', '.export-copy-world-env', () => {
-      Core.copyAsText();
-      this.close();
+    html.on("change", ".toggle-selections", (el) => {
+      $(el.target.closest("fieldset"))
+        .find("td input")
+        .prop("checked", el.target.checked);
     });
 
-    html.on('click', '.export-save-world-settings', () => {
-      Core.exportGameSettings();
-      this.close();
-    });
+    html.on("click", ".import", () => {
+      for (let field of this.form.getElementsByTagName("fieldset")) {
+        let targetType = field.dataset?.type;
+        if (!targetType) {
+          log(false, "Could not find fieldset target type");
+          continue;
+        }
 
-    html.on('change', '.import-world-settings', (el) => {
-      Core.importGameSettings(el.target);
+        switch (targetType) {
+          case "world":
+            this.importWorldSettings(field);
+            break;
+          case "player":
+            this.importPlayerSettings(field);
+            break;
+        }
+      }
+
       this.close();
     });
   }
 
+  importWorldSettings(fieldset) {
+    let changes = [];
+    for (let input of fieldset.elements) {
+      if (!input.checked || !input.name) {
+        continue;
+      }
+
+      let target = input.dataset?.for;
+      if (!this.settings[target]) {
+        log(false, "Import world settings: could not find target for", input);
+        continue;
+      }
+
+      log(false, "Importing world setting", this.settings[target]);
+      changes.push(this.settings[target]);
+    }
+    if (changes.length) {
+      Core.processSettings(changes).then(() => {
+        ui.notifications.info(
+          game.i18n.localize("forien-copy-environment.updatedReloading"),
+          {}
+        );
+        window.setTimeout(window.location.reload.bind(window.location), 5000);
+      });
+    }
+  }
+
+  importPlayerSettings(fieldset) {
+    let targetUser = null;
+    let changes = {
+      flags: {},
+    };
+    for (let input of fieldset.elements) {
+      if (!input.checked || !input.name) {
+        continue;
+      }
+
+      let target = input.dataset?.for;
+      if (!this.playerSettings[target]) {
+        log(false, "Import player settings: could not find target for", input);
+        continue;
+      }
+
+      let type = input.dataset?.type;
+      if (!type) {
+        log(false, "Import player settings: missing type (core or flag)");
+        continue;
+      }
+
+      if (!targetUser) {
+        targetUser = game.users.getName(this.playerSettings[target].name);
+      }
+
+      if (type === "core") {
+        changes[input.name] = this.playerSettings[target].playerDifferences[
+          input.name
+        ].newVal;
+      }
+
+      if (type === "flag") {
+        changes.flags[input.name] = this.playerSettings[
+          target
+        ].playerFlagDifferences[input.name].newVal;
+      }
+    }
+
+    if (!targetUser) {
+      log(false, "No targetUser found.");
+      return;
+    }
+
+    if (Object.keys(changes).length === 1 && isObjectEmpty(changes.flags)) {
+      log(false, "No changes selected for", targetUser?.name);
+      return;
+    }
+
+    log(false, `Updating ${targetUser.name} with`, changes);
+    targetUser.update(changes);
+
+    ui.notifications.info(
+      game.i18n.format("forien-copy-environment.import.updatedPlayer", {
+        name: targetUser.name,
+      }),
+      {}
+    );
+  }
+
   static download(data, filename) {
     if (!filename) {
-      log(false, 'Missing filename on download request');
+      log(false, "Missing filename on download request");
       return;
     }
 
     let jsonStr = JSON.stringify(data, null, 2);
 
-    saveDataToFile(jsonStr, 'application/json', filename);
+    saveDataToFile(jsonStr, "application/json", filename);
   }
 
-  static getData() {
-    let modules = game.data.modules.filter(m => m.active);
+  static data() {
+    let modules = game.data.modules.filter((m) => m.active);
     let system = game.data.system;
     let core = game.data.version;
 
-    let message = "List generated with Forien's Copy Environment: https://github.com/League-of-Foundry-Developers/foundryvtt-forien-copy-environment";
+    let message = game.i18n.localize("forien-copy-environment.message");
 
     return {
       message,
       core,
       system,
-      modules
+      modules,
     };
   }
 
   static getText() {
-    let data = this.getData();
+    let data = this.data();
     let text = `Core Version: ${data.core}\n\n`;
 
     text += `System: ${data.system.id} ${data.system.data.version} (${data.system.data.author}) \n\n`;
 
     text += `Modules: \n`;
-    data.modules.forEach(m => {
+    data.modules.forEach((m) => {
       text += `${m.id} ${m.data.version} (${m.data.author})\n`;
     });
 
@@ -98,91 +247,111 @@ export default class Core extends FormApplication {
   static copyAsText() {
     let text = this.getText();
 
-    const el = document.createElement('textarea');
+    const el = document.createElement("textarea");
     el.value = text;
-    el.setAttribute('readonly', '');
-    el.style.position = 'absolute';
-    el.style.left = '-9999px';
+    el.setAttribute("readonly", "");
+    el.style.position = "absolute";
+    el.style.left = "-9999px";
     document.body.appendChild(el);
     el.select();
-    document.execCommand('copy');
+    document.execCommand("copy");
     document.body.removeChild(el);
 
-    ui.notifications.info("Environment data copied to clipboard!", {});
+    ui.notifications.info(
+      game.i18n.localize("forien-copy-environment.copiedToClipboard"),
+      {}
+    );
   }
 
   static saveSummaryAsJSON() {
-    let data = this.getData();
+    let data = this.data();
 
     data.core = {
-      version: data.core
+      version: data.core,
     };
     data.system = {
       id: data.system.id,
       version: data.system.data.version,
       author: data.system.data.author,
-      manifest: data.system.data.manifest
+      manifest: data.system.data.manifest,
     };
-    data.modules = data.modules.map(m => {
+    data.modules = data.modules.map((m) => {
       return {
         id: m.id,
         version: m.data.version,
         author: m.data.author,
-        manifest: m.data.manifest
-      }
+        manifest: m.data.manifest,
+      };
     });
 
-    this.download(data, 'foundry-environment.json');
+    this.download(data, "foundry-environment.json");
   }
 
   static exportGameSettings() {
-    let data = game.data.settings.map(s => ({
-      key: s.key,
-      value: s.value
-    }));
-    this.download(data, 'foundry-settings-export.json');
+    // Return an array with both the world settings and player settings together.
+    let data = Array.prototype.concat(
+      game.data.settings.map((s) => ({
+        key: s.key,
+        value: s.value,
+      })),
+      game.users.map((u) => ({
+        name: u.data.name,
+        core: {
+          avatar: u.data.avatar,
+          color: u.data.color,
+          permissions: u.data.permissions,
+          role: u.data.role,
+        },
+        flags: u.data.flags,
+      }))
+    );
+    this.download(data, "foundry-settings-export.json");
   }
 
   static importGameSettingsQuick() {
     const input = $('<input type="file">');
-    input.on("change", () => Core.importGameSettings(input.get(0)));
-    input.trigger('click');
+    input.on("change", this.importGameSettings);
+    input.trigger("click");
   }
 
-  static importGameSettings(el) {
-    const file = el?.files[0];
+  static importGameSettings() {
+    const file = this.files[0];
     if (!file) {
-      log(false, 'No file provided for game settings importer.');
+      log(false, "No file provided for game settings importer.");
       return;
     }
 
-    readTextFromFile(file).then(async result => {
-      const settings = JSON.parse(result);
-      log(false, 'import settings', settings);
-      for (const setting of settings) {
-        await Core.processSetting(setting);
+    readTextFromFile(file).then(async (result) => {
+      try {
+        const settings = JSON.parse(result);
+        let coreSettings = new Core(settings);
+        coreSettings.render(true);
+      } catch (e) {
+        log(false, "Could not parse import data.");
       }
-
-      ui.notifications.info("Updated world settings. Reloading world in 5sec...", {});
-      window.setTimeout(window.location.reload.bind(window.location), 5000);
     });
   }
 
-  static async processSetting(setting) {
-    const config = game.settings.settings.get(setting.key);
-    if (config?.scope === "client") {
-      const storage = game.settings.storage.get(config.scope);
-      storage.setItem(setting.key, setting.value);
-    } else if (game.user.isGM) {
-      try {
-        await SocketInterface.dispatch("modifyDocument", {
-          type: "Setting",
-          action: "update",
-          data: setting
-        });
-      } catch (e) {
-        log(false, `Setting key ${setting.key} could not be dispatched to server.`);
-        console.error(e);
+  static async processSettings(settings) {
+    for (const setting of settings) {
+      const config = game.settings.settings.get(setting.key);
+      if (config?.scope === "client") {
+        const storage = game.settings.storage.get(config.scope);
+        storage.setItem(setting.key, setting.value);
+      } else if (game.user.isGM) {
+        try {
+          await SocketInterface.dispatch("modifyDocument", {
+            type: "Setting",
+            action: "update",
+            data: setting,
+          });
+        } catch (e) {
+          log(
+            false,
+            `Setting key ${setting.key} could not be dispatched to server.`
+          );
+          console.error(e);
+        }
       }
     }
   }