+++ scripts/preloadTemplates.js
@@ -1,10 +1,9 @@
-import { templates } from './config.js';
+import { templates } from "./config.js";
 
-export const preloadTemplates = async function() {
-	
-	return loadTemplates(
-		Object.keys(templates).map(function(key){
-    	return templates[key];
-		})
-	);
-}
+export const preloadTemplates = async function () {
+  return loadTemplates(
+    Object.keys(templates).map(function (key) {
+      return templates[key];
+    })
+  );
+};